requirements.txt
asgiref==3.7.2
attrs==24.2.0
blinker==1.4
certifi==2024.8.30
cliapp==1.20180812.1
cmdtest==0.32+git
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
drf-yasg==1.21.8
exceptiongroup==1.2.2
filelock==3.15.1
h11==0.14.0
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.3.6
meson==1.6.0
more-itertools==8.10.0
netifaces==0.11.0
ninja==1.11.1.1
oauthlib==3.2.0
outcome==1.3.0.post0
packaging==24.1
phonenumbers==8.13.50
pillow==11.0.0
platformdirs==4.2.2
pycairo==1.27.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
pytz==2024.2
PyYAML==5.4.1
SecretStorage==3.3.1
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
systemd-python==234
trio==0.26.2
trio-websocket==0.11.1
ttystatus==0.38
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.3
validators==0.34.0
virtualenv==20.26.2
wadllib==1.3.6
websocket-client==1.8.0
wsproto==1.2.0
zipp==1.0.0
